pkg/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["pibragimov <pibragimov@whoosh.bike>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"
sqlalchemy = "^2.0.28"
psycopg2-binary = "^2.9.3"
ujson = "^5.9.0"
fastapi = "^0.110.0"
uvicorn = "^0.23.2"
pyhumps = "^3.8.0"
loguru = "^0.7.2"
asyncpg = "^0.28.0"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
fastapi-cache2 = { extras = ["redis"], version = "^0.2.2" }
ulid-py = "^1.1.0"
geoalchemy2 = "^0.14.1"
aiohttp = "^3.8.6"
requests = "^2.31.0"
pydantic = "v2.6.4"
greenlet = "^3.0.3"
jupyterlab = "^4.1.5"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
